refactor warehouse
