--- frontend/src/app/warehouse-details/warehouse-details.component.html
@@ -1,20 +1,0 @@
-<app-header></app-header>
-<div>
-    <div *ngIf="data != undefined; else noInventory">
-        <h3>Inventory List:</h3>
-        <ul *ngFor="let item of data.inventory">
-            <li>{{item.item.itemId}}</li>
-            <li>{{item.item.name}}</li>
-            <li>{{item.item.description}}</li>
-            <li>{{item.quantity}}</li>
-            <button (click)="deleteWarehouseItem(warehouseId, item.item.itemId ?? 0)">Delete</button>
-            <button (click)="openEditDialog(item)">Edit</button>
-        </ul>
-    </div>
-    <ng-template #noInventory>
-        <h3>No inventory available.</h3>
-    </ng-template>
-    
-    <app-add-warehouse-item-form [warehouseId]="warehouseId" ></app-add-warehouse-item-form>
-</div>
-
