+++ frontend/src/app/warehouse/warehouse-details/component/add-warehouse-item-form/add-warehouse-item-form.component.html
