+++ frontend/src/app/warehouse/warehouse-details/component/add-warehouse-item-form/add-warehouse-item-form.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { WarehouseDetailsService } from '../warehouse-details/warehouse-details.service';
+import { WarehouseDetailsService } from '../../warehouse-details.service';
 
 @Component({
   selector: 'app-add-warehouse-item-form',
