+++ frontend/src/app/warehouse/warehouse-details/component/edit-item-dialog/edit-item-dialog.component.html
