+++ frontend/src/app/warehouse/warehouse-details/warehouse-details.component.html
@@ -1,0 +1,75 @@
+<app-header></app-header>
+<div>
+    <div *ngIf="data != undefined; else noInventory">
+        <h3>Inventory List:</h3>
+        <ul *ngFor="let item of data.inventory">
+            <div class="item">
+                <li>{{item.item.itemId}}</li>
+                <li>{{item.item.name}}</li>
+                <li>{{item.item.description}}</li>
+                <li>{{item.quantity}}</li>
+                <button (click)="deleteWarehouseItem(warehouseId, item.item.itemId ?? 0)">Delete</button>
+                <button (click)="openEditDialog(item)">Edit</button>
+            </div>
+        </ul>
+    </div>
+    <ng-template #noInventory>
+        <h3>No inventory available.</h3>
+    </ng-template>
+    
+    <app-add-warehouse-item-form [warehouseId]="warehouseId" ></app-add-warehouse-item-form>
+</div>
+
+<style>
+    div {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        margin: 20px;
+    }
+
+    h3 {
+        text-align: center;
+        color: #333;
+    }
+
+    ul {
+        list-style-type: none;
+        padding: 0;
+    }
+
+    .item {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        margin: 10px;
+        padding: 15px;
+        border: 1px solid #ccc;
+        border-radius: 8px;
+        background-color: #f9f9f9;
+        box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+        width: 300px;
+    }
+
+    li {
+        margin: 5px 0;
+        font-size: 14px;
+        color: #555;
+    }
+
+    button {
+        margin: 5px;
+        padding: 8px 12px;
+        border: none;
+        border-radius: 4px;
+        background-color: #007bff;
+        color: white;
+        cursor: pointer;
+        font-size: 14px;
+    }
+
+    button:hover {
+        background-color: #0056b3;
+    }
+</style>
