+++ frontend/src/app/warehouse/warehouse-details/warehouse-details.component.spec.ts
