+++ frontend/src/app/warehouse/warehouse-details/warehouse-details.service.spec.ts
